unignore _site during deploys

+++ nix/deploy/default.nix
@@ -21,15 +21,20 @@ if [[ -n $(git status --porcelain) ]]
    echo "Building jekyll docs"
    nix-build
 
-   # if [[ $(git status --porcelain) ]]
-   #  then
+   # we don't want to be ignoring the thing we're supposed to be deploying
+   # if you deploy from local rather than CI you'll need to be careful that you
+   # don't push this change to master on the main repo
+   cat '!_site' >> .gitignore
+
+   if [[ $(git status --porcelain) ]]
+    then
      echo "Pushing to github pages"
      git add . && git commit -am'jekyll build docs [skip ci]'
      # git push ${upstream} ${from-branch}
      git push ${deploy} `git subtree split --prefix ${path} ${from-branch} --message 'github pages push [skip ci]'`:${to-branch} --force
-   #  else
-   #   echo 'failed git status check'
-   # fi
+    else
+     echo 'failed git status check'
+   fi
 fi
   '';
 in
